extract RecordList to separate component
